Move remaining room functions out of pokertable
